-s SOURCE_TYPE -t PACKAGE_TYPE
This introduces the concept of sources and packages.

A source is "something to build a package from" - i.e. something
that can be turned into a versioned tarball.  Easy ones to support
are 'dir' (a plain directory), 'tar' (a tarball), 'npm', 'gem', etc.

A package is a specified output type; deb, rpm, etc. that knows how
to turn a tarball with metadata into a system package.