package names ending with non-version numbers
The regex looks for packages that end with A.B.C
where C is a number i.e. package-0.0.5

This also catches packages that end with a number
but don't denote a version i.e. the python package
httplib2

This commit makes the regex slightly more explicit
by checking for .C not just C