Allow unsetting epoch and iteration by passing empty strings
package.rb initializes epoch and iteration to nil, but the default
arguments for both are '1', which means the package version always
ends up with an epoch and an iteration. By changing the template
to take empty strings into account you can customize the version
string by passing --epoch '' and --iteration ''.