Introduced a dependency mapping for != dependency
There exist gems in the wild that have a dependency !=
Example is sprocket gem (at least version 2.1.3). Use a mapping
that prevents dpkg to choke on != and replaced it with a mapping
>=. May be suboptimal in some cases, though.Introduced a dependency mapping for != dependency

There exist gems in the wild that have a dependency !=
Example is sprocket gem (at least version 2.1.3). Use a mapping
that prevents dpkg to choke on != and replaced it with a mapping
>=. May be suboptimal in some cases, though.Introduced a dependency mapping for != dependency

There exist gems in the wild that have a dependency !=
Example is sprocket gem (at least version 2.1.3). Use a mapping
that prevents dpkg to choke on != and replaced it with a mapping
>=. May be suboptimal in some cases, though.Introduced a dependency mapping for != dependency

There exist gems in the wild that have a dependency !=
Example is sprocket gem (at least version 2.1.3). Use a mapping
that prevents dpkg to choke on != and replaced it with a mapping
>=. May be suboptimal in some cases, though.Introduced a dependency mapping for != dependency

There exist gems in the wild that have a dependency !=
Example is sprocket gem (at least version 2.1.3). Use a mapping
that prevents dpkg to choke on != and replaced it with a mapping
>=. May be suboptimal in some cases, though.Introduced a dependency mapping for != dependency

There exist gems in the wild that have a dependency !=
Example is sprocket gem (at least version 2.1.3). Use a mapping
that prevents dpkg to choke on != and replaced it with a mapping
>=. May be suboptimal in some cases, though.Introduced a dependency mapping for != dependency

There exist gems in the wild that have a dependency !=
Example is sprocket gem (at least version 2.1.3). Use a mapping
that prevents dpkg to choke on != and replaced it with a mapping
>=. May be suboptimal in some cases, though.