python: customize the --install-data option for setup.py
Give the possibility to specify from command line where to install the
data. Data were installed in the 'lib' directory, which is still the
default setting if not explicitely changed, for backward compatibility.
A more "normal" data directory would be /usr or /usr/share, also
depending on the target OS.