Very minor edit. Fixing an incorrectly escaped comment in the deb erb template file. (added '-' to the end tag).
Without the edit, the template was inserting an extra blank line into the
config file. When dpkg/reprepro try to read the file the blank line is
interpreted as the end of the configuration, and the additional lines
below it cause an error.

With the edit, no additional blank line is inserted. The comment is
correctly escaped.

Signed-off-by: Andrew Bunday <andrew.bunday@gmail.com>