pear: add options for bin_dir, php_bin, and php_dir settings
packages with executables will put their bins in bin_dir, which is
prefix/pear by default. php_bin can be used if the php interpreter is
in a different location on the target system. php_dir can be used for
overriding the directory where php files are put (useful on debian).