Fall back to __str__ when encoding package metadata as JSON
This is a fix for finicky packages like unittest2 that pass
non-string parameters for version into setup(). If the JSON module
doesn't know explicitly about the type, it will raise a TypeError.
In this case, we check if the object's class has a __str__ method
defined, and if so, just encode the object as a string.