Add --python-disable-dependency option to disable specific python dependencies.
This is useful for example in the case of a dependency on the python
PyYAML package.

By default, this dependency gets translated to a debian package
dependency on python-pyyaml. That package does not exist on Debian
systems, so a way to override the dependency is needed.

A similar option exists already for gems (--gem-disable-dependency).