virtualenv: add option to also install the package in the virtualenv
one common use-case for a virtualenv requirements.txt source
is to distribute ones software and the dependencies in one
debian file. unfortunately the current implementation doesn't
get one all the way to that goal. Starting with a local checkout
of a python package (containing one's own software, setup.py and
requirements.txt) fpm can create a package that contains the
software dependencies, but not the software itself.

this commit adds an option to also run the package setup.py
after populating the virtualenv, thus giving a means to
package one's own application *and* the dependencies in
one step