Allow suppresing configuration from the source package
The source package may contain undesired metadata. Through the cli we
can only currently set said metadata to an empty string to delete it.

An empty string and nil behave differently for some configuration
options, e.g. iteration results in a hyphen being appended to the
version string like so, 123-

For consistency, if a configuration option is set to an empty string,
store it as nil.