Remove badges from README
Image badges feel like a nice idea, but I'm not sure they add much
value.

The build output has been failing for years, primarily because some
parts of the fpm test suite haven't been maintained or that the actual
test environment on Travis CI is not correct.

The irc link is no longer useful because freenode irc network has been
taken over by an adversarial group.

The gem badge could be informative, but as an image, I wonder if it'd
be better to use text instead of a hotlinked image to rubygems?

Lastly, I'm not sure any of these image badges are accessible to screen
readers.