support a user-specified name mapping
This allows for fpm-generated packages to be made to coexest with
OS-provided packages which have a different name. For example, building
a python module that depends on 'pyyaml' will produce an RPM that
depends on 'python-pyyaml', yet the official redhat/centos package is
named 'PyYAML' and provides 'PyYAML'. An fpm-build package of 'pyyaml',
could provide 'python-pyaml', but installing it would conflict with
'PyYAML'.

Supporting a name mapping allows for this situation to be handled. When
the name map is set, e.g. in this case to '{"python-pyyaml": "PyYAML"}':
* Any dependency on 'python-pyyaml' will be mapped to the target
  official package name of 'PyYAML'.
* Any build of a package that would normally be named 'python-pyyaml' is
  mapped to 'PyYAML' as well, thus providing a drop-in replacement of
  the official package.

Doing this by means of a mapping, rather than by adding/removing
dependencies for the packages that need it, allows for a more simple
automated use. Any build system that makes packages for a target OS can
be configured with a static mapping tailored to that OS, and the mapping
will only be applied to dependencies and package names if and when they
match.