add --cpan-package-reject-from-depends option to filter auto-depends
The auto-depends for cpan modules worked well on rpm because of the
use of capabilities (ie: perl(IO::Handle)) but deb packages of CPAN
modules generate dependencies that might not exist because they are
provided by another package and therefore need to be filtered out.

For instance, perl(IO::Handle) would be turned into libio-handle-perl,
and that package doesn't exist.  The IO/Handle.pm file is actually in
package perl-base.

Prior to this patch, fpm would filter out a short list of
modules (vars, warnings, strict, and Config) and with this patch you
can add to that list with --cpan-package-reject-from-depends.