Bump minimum required version to meet dependencies
While fpm may run on Ruby 1.9, several of it's dependencies no
longer support that release. I recently discovered this trying
to build on an older ruby. The `json` gem requires 2.3 or later
and the `public_suffix` gem requires 2.6 or later. So I'm
proposing this PR to bump the minumum version requirement to
2.6.2 which seems to work without issue.