Add flag --rpm-with-source
This flag allows a user to ask for an SRPM in addition to the binary
rpm.

This also adds fallback case if the file extension seems weird. That is,
if someone asks for a package file named "foobar" then fpm will also
output an SRPM named "foobar.src.rpm" in addition to the rpm file named
"foobar"