CPAN->deb related fixes (#1947)
* fix comment at end of cpan class

* make &perldepfix create correct dependencies for cpan->deb packages

prior version created automatic dependencies for deb packages that
weren't named according to the standard specified here:
https://www.debian.org/doc/packaging-manuals/perl-policy/ch-module_packages.html

* add --cpan-package-name-postfix and fix automatic names cpan pkgs

use --cpan-package-name-prefix and --cpan-package-name-postfix to
create the cpan package name

do not assume a dash between the prefix and the package name

change the defualt prefix from perl to perl- because for deb packages
you want the prefix to be lib and do not want a -

* add --cpan-package-reject-from-depends option to filter auto-depends

The auto-depends for cpan modules worked well on rpm because of the
use of capabilities (ie: perl(IO::Handle)) but deb packages of CPAN
modules generate dependencies that might not exist because they are
provided by another package and therefore need to be filtered out.

For instance, perl(IO::Handle) would be turned into libio-handle-perl,
and that package doesn't exist.  The IO/Handle.pm file is actually in
package perl-base.

Prior to this patch, fpm would filter out a short list of
modules (vars, warnings, strict, and Config) and with this patch you
can add to that list with --cpan-package-reject-from-depends.

* update getting-started.rst to reflect cpan related changes

* Revert "update getting-started.rst to reflect cpan related changes"

This reverts commit ae1a62842129d1d5ca0b8ef94ef7c7ba56a2696e.

* Revert "add --cpan-package-reject-from-depends option to filter auto-depends"

This reverts commit b4e138ffd742cede22c11a2c1126b6f8d23654a7.

* Revert "add --cpan-package-name-postfix and fix automatic names cpan pkgs"

This reverts commit 1564cd9944ef4310392c098042d9b0ebb21a1114.

* add --cpan-package-reject-from-depends option to filter auto-depends

The auto-depends for cpan modules worked well on rpm because of the
use of capabilities (ie: perl(IO::Handle)) but deb packages of CPAN
modules generate dependencies that might not exist because they are
provided by another package and therefore need to be filtered out.

For instance, perl(IO::Handle) would be turned into libio-handle-perl,
and that package doesn't exist.  The IO/Handle.pm file is actually in
package perl-base.

Prior to this patch, fpm would filter out a short list of
modules (vars, warnings, strict, and Config) and with this patch you
can add to that list with --cpan-package-reject-from-depends.

* use Debian style names for CPAN modules

---------

Co-authored-by: Aran Cox <aran-cox@uiowa.edu>
Co-authored-by: Jordan Sissel <131818+jordansissel@users.noreply.github.com>