Fix breakage on Ruby older than v2.6.0
Ruby v2.6.0 introduces "endless range" syntax, such as "foo"[1..] where
`1..` is the endless range.

To fix this for rubies older than 2.6.0, the older syntax can be used,
where `String#[3..]` is the same as `String#[3..-1]` as a negative value
on a range, for a String slice, is relative to the end of the string.

Citations:

https://github.com/ruby/ruby/blob/v2_6_0/NEWS?plain=1#L29-L36
> Endless ranges are introduced

Fixes #2120